Fix loading of `Mock` interface
The mock version of the code attempted to parse a `URI` without checking
it existed first.

So upstream in the main `fog` tests Aliyun was raising a
`URI::InvalidURIError` when the configuration was missing rather than
the correct `ArgumentError`.

The configuration checking code still needs cleaning up but this fixes
the main upstream issues.